pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "agent-firewall"
version = "0.1.0"
description = "Security control plane for AI agent tool use."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
  { name = "Dalton" }
]
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "sqlalchemy[asyncio]>=2.0.36,<3.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "redis[hiredis]>=5.1.0,<6.0.0",
  "opentelemetry-api>=1.28.0,<2.0.0",
  "opentelemetry-sdk>=1.28.0,<2.0.0",
  "opentelemetry-exporter-otlp-proto-http>=1.28.0,<2.0.0",
  "opentelemetry-instrumentation-fastapi>=0.49b0,<1.0.0",
  "opentelemetry-instrumentation-sqlalchemy>=0.49b0,<1.0.0",
  "uvicorn>=0.32.0,<1.0.0",
]

[project.scripts]
agent-firewall-server = "agent_firewall.server:run"

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/agent_firewall"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
